+++ src/components/pokemon/PokemonItemHeader.vue
@@ -101,7 +101,6 @@ export default {
   flex-direction: column;
   align-items: center;
   width: 100%;
-  width: 100%;
   position: relative;
 
   @media (min-width: $min-width-fourth-break) {